refactor layouts
