--- .idea/dictionaries/Zupper.xml
@@ -1,3 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="Zupper" />
-</component>
