+++ app/src/main/java/br/com/estagio/oletrainning/zup/otmovies/confirmInformationActivity/confirmInformationActivity.java
@@ -1,15 +1,15 @@
-package br.com.estagio.oletrainning.zup.otmovies.EnterPasswordActivity;
+package br.com.estagio.oletrainning.zup.otmovies.confirmInformationActivity;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import br.com.estagio.oletrainning.zup.otmovies.R;
 
-public class EnterPasswordActivity extends AppCompatActivity {
+public class confirmInformationActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_enter_password);
+        setContentView(R.layout.activity_confirm_information);
     }
 }
