+++ app/src/main/java/br/com/estagio/oletrainning/zup/otmovies/informTokenAndNewPasswordActivity/informTokenAndNewPasswordActivity.java
@@ -1,0 +1,15 @@
+package br.com.estagio.oletrainning.zup.otmovies.informTokenAndNewPasswordActivity;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+import br.com.estagio.oletrainning.zup.otmovies.R;
+
+public class informTokenAndNewPasswordActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_inform_token_and_new_password);
+    }
+}
