+++ app/src/main/res/layout/activity_inform_token_and_new_password.xml
@@ -9,7 +9,7 @@
 
     <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:id="@+id/layout_finish_registration"
+        android:id="@+id/layout_informTokenAndNewPassword"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:background="@color/colorBackground"
@@ -29,10 +29,10 @@
             android:id="@+id/textView_header"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/marginTopHeaderPreLogin"
+            android:layout_marginTop="@dimen/marginTwentyEight"
             android:fontFamily="@font/dosisbold"
             android:letterSpacing="@dimen/defaultHeaderLetterSpacing"
-            android:text="@string/headerFirst_informTokenAndNewPassword"
+            android:text="@string/token_sended"
             android:textAlignment="center"
             android:textColor="@color/colorEmailHeadHint"
             android:textSize="@dimen/textSizeHeader"
@@ -59,13 +59,13 @@
             android:id="@+id/errorEditText_token"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/marginTopErrorEditText_token"
+            android:layout_marginTop="@dimen/bigMargin"
             android:layout_marginBottom="@dimen/marginBottonErrorEditText_token"
             app:layout_constraintBottom_toTopOf="@id/button_nextPreLogin"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/textView_emailEntered"
-            app:setHint="@string/code"
+            app:setHint="@string/tokenValidation"
             app:setInicialErrorVisibility="false"
             app:setInputType="text"
             app:setMessageError="@string/errorCode">
@@ -100,13 +100,64 @@
             app:layout_constraintStart_toEndOf="@+id/textView_NotReceiveToken"
             app:layout_constraintTop_toBottomOf="@+id/errorEditText_token" />
 
+        <TextView
+            android:id="@+id/textView_header_newPassword"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/marginStandard"
+            android:layout_marginTop="@dimen/marginTopMedium"
+            android:layout_marginEnd="@dimen/marginStandard"
+            android:fontFamily="@font/dosisbold"
+            android:letterSpacing="@dimen/defaultHeaderLetterSpacing"
+            android:text="@string/create_new_password"
+            android:textAlignment="center"
+            android:textColor="@color/colorEmailHeadHint"
+            android:textSize="@dimen/textSizeHeader"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/errorEditText_token" />
+
+        <br.com.estagio.oletrainning.zup.otmovies.CustomComponents.ComponentErrorEditText
+            android:id="@+id/errorEditText_password"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/marginStandard"
+            android:layout_marginTop="@dimen/marginTopSmall"
+            android:layout_marginEnd="@dimen/marginStandard"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/errorEditText_confirmPassword"
+            app:setHint="@string/newPassword"
+            app:setInicialErrorVisibility="false"
+            app:setInputType="textPassword"
+            app:setMessageError="@string/error_password">
+
+        </br.com.estagio.oletrainning.zup.otmovies.CustomComponents.ComponentErrorEditText>
+
+        <br.com.estagio.oletrainning.zup.otmovies.CustomComponents.ComponentErrorEditText
+            android:id="@+id/errorEditText_confirmPassword"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="@dimen/marginStandard"
+            android:layout_marginTop="@dimen/bigMargin"
+            android:layout_marginEnd="@dimen/marginStandard"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/textView_header_newPassword"
+            app:setHint="@string/passwordConfirm"
+            app:setInicialErrorVisibility="false"
+            app:setInputType="textPassword"
+            app:setMessageError="@string/noMatchPassword">
+
+        </br.com.estagio.oletrainning.zup.otmovies.CustomComponents.ComponentErrorEditText>
+
         <Button
-            android:id="@+id/button_validate"
+            android:id="@+id/button_changePassword"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@drawable/button_style"
             android:fontFamily="@font/dosisbold"
-            android:text="@string/validate"
+            android:text="@string/change_password"
             android:textColor="@color/colorTextBottonEmailPreLogin"
             android:textSize="@dimen/textSizeDefault"
             app:layout_constraintBottom_toBottomOf="parent"
