+++ app/src/main/res/layout/activity_login.xml
@@ -16,10 +16,11 @@
         android:background="@color/colorBackground">
 
         <ImageView
+            android:contentDescription="@string/imageview_backarrow"
             android:id="@+id/imageView_backArrow"
             android:layout_width="@dimen/sizeBackArrow"
             android:layout_height="@dimen/sizeBackArrow"
-            android:layout_margin="30dp"
+            android:layout_margin="@dimen/marginThirty"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/linearLayout_red_toast"
             app:srcCompat="@drawable/ic_arrow_back_black_24dp" />
@@ -28,17 +29,16 @@
             android:id="@+id/textView_header"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="76dp"
-            android:layout_marginLeft="@dimen/paddingLayout"
+            android:layout_marginTop="@dimen/marginSeventySix"
+            android:layout_marginStart="@dimen/paddingLayout"
             android:layout_marginEnd="@dimen/paddingLayout"
             android:fontFamily="@font/dosisbold"
             android:letterSpacing="@dimen/defaultHeaderLetterSpacing"
-            android:text="INFORME SUA SENHA"
+            android:text="@string/enterPassword"
             android:textAlignment="center"
             android:textColor="@color/colorEmailHeadHint"
             android:textSize="@dimen/textSizeHeader"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.559"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/linearLayout_red_toast" />
 
@@ -46,7 +46,7 @@
             android:id="@+id/textView_emailEntered"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
+            android:layout_marginBottom="@dimen/bigMargin"
             android:fontFamily="@font/robotoregular"
             android:text="@string/userEmailEntered"
             android:textAlignment="center"
@@ -59,15 +59,15 @@
             android:id="@+id/errorEditText_login"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_margin="30dp"
+            android:layout_margin="@dimen/marginThirty"
             app:layout_constraintBottom_toTopOf="@+id/button_sign_in"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/textView_header"
-            app:setHint="Senha"
+            app:setHint="@string/password"
             app:setInicialErrorVisibility="false"
             app:setInputType="text"
-            app:setMessageError="Código inválido">
+            app:setMessageError="@string/errorCode">
 
         </br.com.estagio.oletrainning.zup.otmovies.CustomComponents.ComponentErrorEditText>
 
@@ -75,10 +75,10 @@
             android:id="@+id/textView_password"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="8dp"
+            android:layout_marginTop="@dimen/marginDefault"
+            android:layout_marginEnd="@dimen/marginDefault"
             android:fontFamily="@font/robotoregular"
-            android:text="Esqueceu sua senha?"
+            android:text="@string/forgotPassword"
             android:textAlignment="center"
             android:textColor="@color/colorTextDefault"
             android:textSize="@dimen/textSizeDefault"
@@ -92,7 +92,7 @@
             android:background="@drawable/button_style"
             android:fontFamily="@font/dosisbold"
             android:layout_margin="@dimen/paddingLayout"
-            android:text="entrar"
+            android:text="@string/enter"
             android:textColor="@color/colorTextBottonEmailPreLogin"
             android:textSize="@dimen/textSizeDefault"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -102,7 +102,7 @@
         <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
             android:id="@+id/linearLayout_red_toast"
             android:layout_width="match_parent"
-            android:layout_height="50dp"
+            android:layout_height="@dimen/toastHeight"
             android:visibility="visible"
             android:background="@color/colorBorderEmailInputError"
             android:orientation="vertical">
@@ -112,14 +112,14 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_gravity="center_horizontal"
-                android:layout_margin="16dp"
-                android:layout_marginStart="8dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginEnd="8dp"
-                android:layout_marginBottom="8dp"
+                android:layout_margin="@dimen/bigMargin"
+                android:layout_marginStart="@dimen/marginDefault"
+                android:layout_marginTop="@dimen/marginDefault"
+                android:layout_marginEnd="@dimen/marginDefault"
+                android:layout_marginBottom="@dimen/marginDefault"
                 android:fontFamily="@font/robotoregular"
                 android:gravity="center_vertical"
-                android:text="Usuário ou senha inválido!"
+                android:text="@string/errorUserOrPassword"
                 android:textColor="@color/colorTextWhite"
                 android:textSize="@dimen/textSizeDefault" />
