+++ app/src/main/res/layout/component_error_edit_text.xml
@@ -9,6 +9,8 @@
     android:orientation="vertical">
 
     <EditText
+        android:importantForAutofill="no"
+        android:inputType="text"
         android:id="@+id/editText_ErrorEditText"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -27,8 +29,8 @@
         android:layout_height="wrap_content"
         android:layout_alignBottom="@id/editText_ErrorEditText"
         android:layout_marginStart="@dimen/marginDefault"
-        android:layout_marginTop="8sp"
-        android:layout_marginEnd="8dp"
+        android:layout_marginTop="@dimen/marginDefault"
+        android:layout_marginEnd="@dimen/marginDefault"
         android:drawableEnd="@drawable/ic_warning_black_24dp"
         android:fontFamily="@font/robotoregular"
         android:text="@string/error_email_pre_login"
