+++ app/src/main/res/layout/component_red_toast.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/toast_layout_root"
     android:layout_width="match_parent"
-    android:layout_height="50dp"
+    android:layout_height="@dimen/toastHeight"
     android:background="@color/colorBorderEmailInputError"
     android:orientation="vertical">
 
@@ -12,13 +12,13 @@
         android:id="@+id/textView_red_toast"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_margin="16dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
+        android:layout_margin="@dimen/bigMargin"
+        android:layout_marginStart="@dimen/marginDefault"
+        android:layout_marginTop="@dimen/marginDefault"
+        android:layout_marginEnd="@dimen/marginDefault"
+        android:layout_marginBottom="@dimen/marginDefault"
         android:fontFamily="@font/robotoregular"
-        android:text="Usuário ou senha inválido!"
+        android:text="@string/errorUserOrPassword"
         android:textColor="@color/colorTextWhite"
         android:textSize="@dimen/textSizeDefault"
         android:gravity="center_vertical"
